whisper-diarization-no/requirements.txt
torch
requests
pyannote.audio>=3.3.1
faster-whisper>=1.0.3
torchvision>=0.15.2
torchtext>=0.15.2
fastapi
uvicorn
faster-whisper
pyannote.audio
torchaudio
aws-xray-sdk
aiohttp
wheel